refactor: rename testing-utils paths to portable-stories
