+++ code/lib/preview-api/src/modules/store/csf/index.ts
@@ -7,4 +7,4 @@ export * from './normalizeProjectAnnotations';
 export * from './getValuesFromArgTypes';
 export * from './composeConfigs';
 export * from './stepRunners';
-export * from './testing-utils';
+export * from './portable-stories';
