+++ code/renderers/react/src/index.ts
@@ -4,7 +4,7 @@ import './globals';
 
 export * from './public-types';
 
-export * from './testing-api';
+export * from './portable-stories';
 
 // optimization: stop HMR propagation in webpack
 if (typeof module !== 'undefined') module?.hot?.decline();
