+++ code/renderers/vue3/src/__tests__/composeStories/composeStories.test.ts
@@ -5,7 +5,7 @@ import { expectTypeOf } from 'expect-type';
 import type { Meta } from '@storybook/vue3';
 import * as stories from './Button.stories';
 import type Button from './Button.vue';
-import { composeStories, composeStory, setProjectAnnotations } from '../../testing-api';
+import { composeStories, composeStory, setProjectAnnotations } from '../../portable-stories';
 
 // example with composeStories, returns an object with all stories composed with args/decorators
 const { CSF3Primary } = composeStories(stories);
